修正一些在msvc下的错误

+++ src/signAChart/QwtPlotItemDataModel.cpp
@@ -465,7 +465,7 @@ double QwtPlotItemDataModel::getItemData(int row, int col, QwtPlotItem *item) co
     case QwtPlotItem::Rtti_PlotCurve:
     {
         const QwtPlotCurve* p = static_cast<const QwtPlotCurve*>(item);
-        if(row < p->dataSize())
+        if((size_t)row < p->dataSize())
         {
             switch(col)
             {
@@ -478,7 +478,7 @@ double QwtPlotItemDataModel::getItemData(int row, int col, QwtPlotItem *item) co
     case QwtPlotItem::Rtti_PlotBarChart:
     {
         const QwtPlotBarChart* p = static_cast<const QwtPlotBarChart*>(item);
-        if(row < p->dataSize())
+        if((size_t)row < p->dataSize())
         {
             switch(col)
             {
@@ -491,7 +491,7 @@ double QwtPlotItemDataModel::getItemData(int row, int col, QwtPlotItem *item) co
     case QwtPlotItem::Rtti_PlotSpectroCurve:
     {
         const QwtPlotSpectroCurve* p = static_cast<const QwtPlotSpectroCurve*>(item);
-        if(row < p->dataSize())
+        if((size_t)row < p->dataSize())
         {
             switch(col)
             {
@@ -505,7 +505,7 @@ double QwtPlotItemDataModel::getItemData(int row, int col, QwtPlotItem *item) co
     case QwtPlotItem::Rtti_PlotIntervalCurve:
     {
         const QwtPlotIntervalCurve* p = static_cast<const QwtPlotIntervalCurve*>(item);
-        if(row < p->dataSize())
+        if((size_t)row < p->dataSize())
         {
             switch(col)
             {
@@ -519,7 +519,7 @@ double QwtPlotItemDataModel::getItemData(int row, int col, QwtPlotItem *item) co
     case QwtPlotItem::Rtti_PlotHistogram:
     {
         const QwtPlotHistogram* p = static_cast<const QwtPlotHistogram*>(item);
-        if(row < p->dataSize())
+        if((size_t)row < p->dataSize())
         {
             switch(col)
             {
@@ -533,7 +533,7 @@ double QwtPlotItemDataModel::getItemData(int row, int col, QwtPlotItem *item) co
     case QwtPlotItem::Rtti_PlotTradingCurve:
     {
         const QwtPlotTradingCurve* p = static_cast<const QwtPlotTradingCurve*>(item);
-        if(row < p->dataSize())
+        if((size_t)row < p->dataSize())
         {
             switch(col)
             {
@@ -549,7 +549,7 @@ double QwtPlotItemDataModel::getItemData(int row, int col, QwtPlotItem *item) co
     case QwtPlotItem::Rtti_PlotMultiBarChart:
     {
         const QwtPlotMultiBarChart* p = static_cast<const QwtPlotMultiBarChart*>(item);
-        if(row < p->dataSize())
+        if((size_t)row < p->dataSize())
         {
             const QwtSetSample& s = p->sample(row);
             if(0 == col)
